feat(entity): 更新实体类以支持MyBatis Plus注解和数据校验
- 为Appointment、ChatRecord、Department、Doctor、MedicalRecord、MessageQueueLog、Patient、PatientSummary、Session、SymptomRecord等实体类添加MyBatis Plus相关注解
- 添加jakarta.validation.constraints.NotBlank注解用于字段校验
- 更新字段类型从int到Long以匹配数据库设计
- 为Schedule实体类统一字段类型并添加TableName注解
- 修改deductStock方法参数类型为Long以保持一致性
- 更新StartSessionTool和SymptomExtractTool中的参数类型及枚举使用方式

+++ src/main/java/com/dd/ai_medical_triage/tool/SymptomExtractTool.java
@@ -16,7 +16,7 @@ public class SymptomExtractTool {
     private SymptomRecordService symptomRecordService;
 
     @Tool(description = "记录患者的症状")
-    public ResultVO extractSymptoms(@ToolParam(description = "患者病历号") int caseNumber,
+    public ResultVO extractSymptoms(@ToolParam(description = "患者病历号") Long caseNumber,
                                     @ToolParam(description = "患者症状") String symptoms,
                                     @ToolParam(description = "症状的程度") String severity,
                                     @ToolParam(description = "持续时间") String duration) {
